refactor(renderer): split Template.Renderer into RenderNode and RenderProp for better modularity; update cabal and fix all type issues. All tests but one pass (function call on model value edge case).
